further refactor based on feedback
