+++ src/index.js
@@ -1,2 +1,2 @@
-export { MediaServer } from './media-server.js'
+export { FastifyController as MediaServer } from './fastify-controller.js'
 export { MapeoManager } from './mapeo-manager.js'
