+++ test-e2e/core-ownership.js
@@ -1,16 +1,17 @@
 import { test } from 'brittle'
 import { KeyManager } from '@mapeo/crypto'
-import { MapeoManager } from '../src/mapeo-manager.js'
-import { kCoreOwnership } from '../src/mapeo-project.js'
 import { parseVersionId } from '@mapeo/schema'
 import RAM from 'random-access-memory'
 import { discoveryKey } from 'hypercore-crypto'
-import { MediaServer } from '../src/media-server.js'
+import Fastify from 'fastify'
+
+import { kCoreOwnership } from '../src/mapeo-project.js'
+import { MapeoManager } from '../src/mapeo-manager.js'
 
 test('CoreOwnership', async (t) => {
   const rootKey = KeyManager.generateRootKey()
   const km = new KeyManager(rootKey)
-  const mediaServer = new MediaServer()
+  const fastify = Fastify()
   const manager = new MapeoManager({
     rootKey,
     projectMigrationsFolder: new URL('../drizzle/project', import.meta.url)
@@ -19,7 +20,7 @@ test('CoreOwnership', async (t) => {
       .pathname,
     dbFolder: ':memory:',
     coreStorage: () => new RAM(),
-    mediaServer,
+    fastify,
   })
 
   const projectId = await manager.createProject()
