+++ test-e2e/manager-basic.js
@@ -4,7 +4,7 @@ import { randomBytes, createHash } from 'crypto'
 import { KeyManager } from '@mapeo/crypto'
 import RAM from 'random-access-memory'
 import { MapeoManager } from '../src/mapeo-manager.js'
-import { MediaServer } from '../src/media-server.js'
+import Fastify from 'fastify'
 
 const projectMigrationsFolder = new URL('../drizzle/project', import.meta.url)
   .pathname
@@ -12,15 +12,13 @@ const clientMigrationsFolder = new URL('../drizzle/client', import.meta.url)
   .pathname
 
 test('Managing created projects', async (t) => {
-  const mediaServer = new MediaServer()
-
   const manager = new MapeoManager({
     rootKey: KeyManager.generateRootKey(),
     projectMigrationsFolder,
     clientMigrationsFolder,
     dbFolder: ':memory:',
     coreStorage: () => new RAM(),
-    mediaServer,
+    fastify: Fastify(),
   })
 
   const project1Id = await manager.createProject()
@@ -119,14 +117,13 @@ test('Managing created projects', async (t) => {
 })
 
 test('Managing added projects', async (t) => {
-  const mediaServer = new MediaServer()
   const manager = new MapeoManager({
     rootKey: KeyManager.generateRootKey(),
     projectMigrationsFolder,
     clientMigrationsFolder,
     dbFolder: ':memory:',
     coreStorage: () => new RAM(),
-    mediaServer,
+    fastify: Fastify(),
   })
 
   const project1Id = await manager.addProject(
@@ -195,14 +192,13 @@ test('Managing added projects', async (t) => {
 })
 
 test('Managing both created and added projects', async (t) => {
-  const mediaServer = new MediaServer()
   const manager = new MapeoManager({
     rootKey: KeyManager.generateRootKey(),
     projectMigrationsFolder,
     clientMigrationsFolder,
     dbFolder: ':memory:',
     coreStorage: () => new RAM(),
-    mediaServer,
+    fastify: Fastify(),
   })
 
   const createdProjectId = await manager.createProject({
@@ -240,14 +236,13 @@ test('Managing both created and added projects', async (t) => {
 })
 
 test('Manager cannot add project that already exists', async (t) => {
-  const mediaServer = new MediaServer()
   const manager = new MapeoManager({
     rootKey: KeyManager.generateRootKey(),
     projectMigrationsFolder,
     clientMigrationsFolder,
     dbFolder: ':memory:',
     coreStorage: () => new RAM(),
-    mediaServer,
+    fastify: Fastify(),
   })
 
   const existingProjectId = await manager.createProject()
@@ -269,8 +264,6 @@ test('Manager cannot add project that already exists', async (t) => {
 })
 
 test('Consistent storage folders', async (t) => {
-  const mediaServer = new MediaServer()
-
   /** @type {string[]} */
   const storageNames = []
   const manager = new MapeoManager({
@@ -278,7 +271,7 @@ test('Consistent storage folders', async (t) => {
     projectMigrationsFolder,
     clientMigrationsFolder,
     dbFolder: ':memory:',
-    mediaServer,
+    fastify: Fastify(),
     coreStorage: (name) => {
       storageNames.push(name)
       return new RAM()
