+++ test-e2e/manager-fastify-server.js
@@ -8,9 +8,10 @@ import FakeTimers from '@sinonjs/fake-timers'
 import { Agent, fetch as uFetch } from 'undici'
 import fs from 'fs/promises'
 import RAM from 'random-access-memory'
+import Fastify from 'fastify'
 
 import { MapeoManager } from '../src/mapeo-manager.js'
-import { MediaServer } from '../src/media-server.js'
+import { FastifyController } from '../src/fastify-controller.js'
 
 const BLOB_FIXTURES_DIR = fileURLToPath(
   new URL('../tests/fixtures/blob-api/', import.meta.url)
@@ -22,7 +23,8 @@ const clientMigrationsFolder = new URL('../drizzle/client', import.meta.url)
   .pathname
 
 test('start/stop lifecycle', async (t) => {
-  const mediaServer = new MediaServer()
+  const fastify = Fastify()
+  const fastifyController = new FastifyController({ fastify })
 
   const manager = new MapeoManager({
     rootKey: KeyManager.generateRootKey(),
@@ -30,12 +32,13 @@ test('start/stop lifecycle', async (t) => {
     clientMigrationsFolder,
     dbFolder: ':memory:',
     coreStorage: () => new RAM(),
-    mediaServer,
+    fastify,
   })
 
-  const project = await manager.getProject(await manager.createProject())
+  // Manager should await for the server to start internally
+  fastifyController.start()
 
-  await manager.startMediaServer()
+  const project = await manager.getProject(await manager.createProject())
 
   const blobUrl1 = await project.$blobs.getUrl({
     driveId: randomBytes(32).toString('hex'),
@@ -45,8 +48,6 @@ test('start/stop lifecycle', async (t) => {
   })
   const response1 = await fetch(blobUrl1)
   t.is(response1.status, 404, 'server started and listening')
-
-  await manager.startMediaServer()
 
   const blobUrl2 = await project.$blobs.getUrl({
     driveId: randomBytes(32).toString('hex'),
@@ -60,13 +61,14 @@ test('start/stop lifecycle', async (t) => {
     'server port is the same'
   )
 
-  await manager.stopMediaServer()
+  await fastifyController.stop()
 
   await t.exception.all(async () => {
     await fetch(blobUrl2)
   }, 'failed to fetch due to connection error')
 
-  await manager.startMediaServer()
+  // Manager should await for the server to start internally
+  fastifyController.start()
 
   const blobUrl3 = await project.$blobs.getUrl({
     driveId: randomBytes(32).toString('hex'),
@@ -77,7 +79,7 @@ test('start/stop lifecycle', async (t) => {
   const response3 = await fetch(blobUrl3)
   t.is(response3.status, 404, 'server started and listening')
 
-  await manager.stopMediaServer()
+  await fastifyController.stop()
 
   await t.exception.all(async () => {
     await fetch(blobUrl3)
@@ -88,14 +90,15 @@ test('retrieving blobs using url', async (t) => {
   const clock = FakeTimers.install({ shouldAdvanceTime: true })
   t.teardown(() => clock.uninstall())
 
-  const mediaServer = new MediaServer()
+  const fastify = Fastify()
+  const fastifyController = new FastifyController({ fastify })
   const manager = new MapeoManager({
     rootKey: KeyManager.generateRootKey(),
     projectMigrationsFolder,
     clientMigrationsFolder,
     dbFolder: ':memory:',
     coreStorage: () => new RAM(),
-    mediaServer,
+    fastify,
   })
 
   const project = await manager.getProject(await manager.createProject())
@@ -112,7 +115,8 @@ test('retrieving blobs using url', async (t) => {
   clock.tick(100_000)
   await exceptionPromise1
 
-  await manager.startMediaServer()
+  // Manager should await for the server to start internally
+  fastifyController.start()
 
   await t.test('blob does not exist', async (st) => {
     const blobUrl = await project.$blobs.getUrl({
@@ -163,7 +167,7 @@ test('retrieving blobs using url', async (t) => {
     st.alike(body, expected, 'matching reponse body')
   })
 
-  await manager.stopMediaServer()
+  await fastifyController.stop()
 
   const exceptionPromise2 = t.exception(async () => {
     await project.$blobs.getUrl({
@@ -181,14 +185,15 @@ test('retrieving icons using url', async (t) => {
   const clock = FakeTimers.install({ shouldAdvanceTime: true })
   t.teardown(() => clock.uninstall())
 
-  const mediaServer = new MediaServer()
+  const fastify = Fastify()
+  const fastifyController = new FastifyController({ fastify })
   const manager = new MapeoManager({
     rootKey: KeyManager.generateRootKey(),
     projectMigrationsFolder,
     clientMigrationsFolder,
     dbFolder: ':memory:',
     coreStorage: () => new RAM(),
-    mediaServer,
+    fastify,
   })
 
   const project = await manager.getProject(await manager.createProject())
@@ -204,7 +209,7 @@ test('retrieving icons using url', async (t) => {
   clock.tick(100_000)
   await exceptionPromise1
 
-  await manager.startMediaServer()
+  await fastifyController.start()
 
   await t.test('icon does not exist', async (st) => {
     const nonExistentIconId = randomBytes(32).toString('hex')
@@ -263,7 +268,7 @@ test('retrieving icons using url', async (t) => {
     st.alike(body, iconBuffer, 'matching response body')
   })
 
-  await manager.stopMediaServer()
+  await fastifyController.stop()
 
   const exceptionPromise2 = t.exception(async () => {
     await project.$icons.getIconUrl(randomBytes(32).toString('hex'), {
